% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COX.R
\name{COX}
\alias{COX}
\title{Function to implement proportional hazard model}
\usage{
COX(form, dat, trControl = NULL)
}
\arguments{
\item{form}{survival formula}

\item{dat}{data frame}

\item{trControl}{list of control parameters:
 \enumerate{
 \item maxit: number of iterations in \code{glmnet}  
 \item regularize: train regularize cox?
 \item nfolds :  number of folds in \code{glmnet} 
 \item lambda : numeric vector of lambda values in \code{glmnet} 
 \item alpha : numeric vector of alpha values in \code{cva.glmnet}    
}}
}
\value{
returns a list with items: 
\itemize{
\item{model: }{ object of class coxph}
\item{form: }{ survival formula} 
}
}
\description{
Train the  Cox model. Optionally, the regularize cox can also be trained based on the implementation in \code{glmnetUtils}.
}
