% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{denormalize}
\alias{denormalize}
\title{Denormalized data}
\usage{
denormalize(normalized, min, max)
}
\arguments{
\item{min}{minimum value of each variable in the original data. This value is stored as an attribute of \code{normalize}}

\item{max}{maximum value of each variable in the original data. This value is stored as an attribute of \code{normalize}}
}
\value{
Original un-normalized data
}
\description{
Take the ouput of \code{normalize} and convert back to original scale.
}
