% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{getResults.ci2}
\alias{getResults.ci2}
\title{get Summary Results 2}
\usage{
getResults.ci2(tab, alpha = 0.05, groups = c("model", "status"),
  stats = c("PCC", "AUC", "sensitivity", "specificity", "G.mean", "BER",
  "Pos Pred Value"))
}
\arguments{
\item{tab}{table with performance results}

\item{alpha}{confidence level}

\item{groups}{variable in \code{tab} to group by}
}
\value{
data frame with summaries (confidence interavals are represented in brackets)
}
\description{
Takes a table of performance metrics, such as cross-validation results and compute summaries (mean and confidence interval) 
ready for publication.
}
