% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLSurvival.R
\name{MLSurvival}
\alias{MLSurvival}
\title{MLSurvival}
\usage{
MLSurvival(form, dat, method, predict.times, trControl, parallel = FALSE,
  dummy.vars = TRUE, mc.cores = 2, seed = 123, ...)
}
\arguments{
\item{form}{survival formula}

\item{dat}{data frame}

\item{method}{character verctor of machine learning algorithms. Implemented algorithms 
 \enumerate{
 \item glm  logistic regression 
 \item glmnet elastic net 
 \item gbm gradient boosting machine 
 \item ranger random forest 
 \item svmRadial  support vector machine with radial basis kernel 
 \item xgbTree extreem boosting machine  
}}

\item{predict.times}{numeric vector containing the survival prediction times}

\item{trControl}{list of control parameters for caret and the ranger models}

\item{parallel}{run cross-validation in parallel?}

\item{dummy.vars}{create  dummy variables/model.matrix}

\item{mc.cores}{number of cores}

\item{seed}{random seed}

\item{\dots}{further arguments passed to caret or other methods.}
}
\value{
returns a list with items: 
\itemize{
\item{model:}{ trained survival model}
\item{perf:}{ performance of models at each survival prediction time: PCC, AUC, sensitivity, specificity, g-mean etc. } 
\item{perf.ave:}{ average of perf with confidence intervals} 
}
}
\description{
Train and evelaute machine learning survival and classification models for time to event data
}
