% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{normalize}
\alias{normalize}
\title{Normalize data}
\usage{
normalize(x)
}
\arguments{
\item{x}{data frame}
}
\value{
Normalized data with attributes min and max representing the min and max of each variable in \code{x}
}
\description{
Normalize data to (0,1)
}
