% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{opt.thresh}
\alias{opt.thresh}
\title{Optimal Threshold}
\usage{
opt.thresh(prob, obs, opt.methods = 9)
}
\arguments{
\item{prob}{predicted probabilities}

\item{obs}{binary (0-1) ground truth}

\item{opt.methods}{optimal threshold method. See \code{Presence.Absence} package}
}
\value{
optimal threshold
}
\description{
Compute the optimal classifcation threshold based on the \code{optimal.thresholds} function in the 
\code{Presence.Absence} package
}
