% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{Performance.measures}
\alias{Performance.measures}
\title{Performance metrics}
\usage{
Performance.measures(pred, obs, threshold = NULL)
}
\arguments{
\item{pred}{predicted probabilities}

\item{obs}{binary (0-1) ground truth}

\item{threshold}{optimal threshold.}
}
\value{
A data frame with performance metrics.
}
\description{
Compute several performance metrics
}
