% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.MLSurvival}
\alias{predict.MLSurvival}
\title{Get Predictions for new data}
\usage{
\method{predict}{MLSurvival}(object, method = "gbm", newdata,
  predict.times)
}
\arguments{
\item{newdata}{new data to get predictions}

\item{oject}{oject}
}
\description{
Get Predictions for new data
}
