% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{predictSurvProb_Cox}
\alias{predictSurvProb_Cox}
\title{Predict Cox}
\usage{
predictSurvProb_Cox(object, newdata, times)
}
\arguments{
\item{object}{trained cox model. Output of \code{train_Cox}}

\item{newdata}{out of sample data}

\item{times}{new time points}
}
\value{
predicted probabilities at eact time point in \code{times}
}
\description{
Get predicted probabilities from the cox model for new data at different time points
}
