% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{predictSurvProb_ranger}
\alias{predictSurvProb_ranger}
\title{Predict Ranger}
\usage{
predictSurvProb_ranger(object, newdata, times, ...)
}
\arguments{
\item{object}{trained ranger model. Output of \code{train_ranger}}

\item{newdata}{out of sample data}

\item{times}{new time points}

\item{\dots}{further arguments passed to caret or other methods.}
}
\value{
predicted probabilities at eact time point in \code{times}
}
\description{
Get predicted probabilities from the ranger model for new data at different time points
}
