% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.classifier.R
\name{train.classifier}
\alias{train.classifier}
\alias{train_classifier}
\title{Predict Survival with classifcation methods}
\usage{
train_classifier(form, dat, method = "gbm", predict.times,
  trControl = NULL, parallel = FALSE, mc.cores = 2, seed = 123,
  ...)
}
\arguments{
\item{form}{survival formula}

\item{dat}{data frame}

\item{method}{classifcation algorithm. The following algorithms have been implemented. 
 \enumerate{
 \item glm  logistic regression 
 \item glmnet elastic net 
 \item gbm gradient boosting machine 
 \item ranger random forest 
 \item svmRadial  support vector machine with radial basis kernel 
 \item xgbTree extreme gradient boosting machine  
}}

\item{predict.times}{survival prediction times}

\item{trControl}{control parameters for the caret train function. Set to NULL to use a default 5-fold cross-validation}

\item{parallel}{run cross-validation in parallel? Uses mclapply which works only on linux}

\item{\dots}{further arguments passed to caret or other methods.}

\item{tuneLength}{same as tuneLength in the caret package}
}
\value{
returns a list with items: 
\itemize{
\item{finalModel: }{ final model trained on the complete data (dat) using optimal tuning paramters}
\item{fitted: }{ predictions on complete data (dat)} 
\item{threshold: }{ optimal classification threshold}
\item{resamples: }{ cross-validation results: predictions on resampled data } 
\item{predict.times: }{ survival prediction times}
\item{bestTune: }{ optimal tuning parameters} 
\item{method: }{ classification algorithm}
}
}
\description{
Train  machine learning classification models on time to event data using the caret package
}
