% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{VimPlot}
\alias{VimPlot}
\title{Plot variable importance}
\usage{
VimPlot(x, top = min(20, length(x$importance)), ...)
}
\arguments{
\item{x}{data frame with variable importance}

\item{top}{number of variables to plot}
}
\description{
Plot variable importance
}
