% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.cox.R
\name{cox}
\alias{cox}
\alias{train_cox}
\title{Cox proportional hazard model}
\usage{
train_cox(form, dat, predict.times, trControl = NULL, parallel = FALSE,
  mc.cores = 2, seed = 123, ...)
}
\arguments{
\item{form}{survival formula}

\item{dat}{data frame}

\item{predict.times}{survival prediction times}

\item{trControl}{list of control parameters:
 \enumerate{
 \item number: number of cross-validations 
 \item regularize: train regularize cox?   
}}

\item{parallel}{run cross-validation in parallel? Uses mclapply which works only on linux}

\item{\dots}{further arguments passed to caret or other methods.}

\item{tuneLength}{same as tuneLength in the caret package}
}
\value{
returns a list with items: 
\itemize{
\item{finalModel: }{ final model trained on the complete data (dat) using optimal tuning paramters}
\item{fitted: }{ predictions on complete data (dat)} 
\item{threshold: }{ optimal classification threshold}
\item{resamples: }{ cross-validation results: predictions on resampled data } 
\item{predict.times: }{ survival prediction times}
\item{bestTune: }{ optimal tuning parameters} 
}
}
\description{
Train the Cox model through cross-validation and select the optimal survival classification threshold. 
A regularized Cox approach which performs feature selection is also implemeted
}
\author{
Che Ngufor <Ngufor.Che@mayo.edu>
}
